Saving progress

+++ src/main/java/nl/hu/cisq1/lingo/trainer/data/SpringGameRepository.java
@@ -3,6 +3,6 @@ package nl.hu.cisq1.lingo.trainer.data;
 import nl.hu.cisq1.lingo.trainer.domain.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SpringGameRepository extends JpaRepository<Game, Long> {
+public interface SpringGameRepository extends JpaRepository<Game, Integer> {
 
 }
